requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cycler==0.11.0
Cython==0.29.28
exceptiongroup==1.0.4
fonttools==4.38.0
gensim==4.2.0
h11==0.14.0
idna==3.4
joblib==1.2.0
JPype1==1.4.1
kiwisolver==1.4.4
konlpy==0.6.0
lxml==4.9.1
matplotlib==3.5.3
numpy==1.21.6
outcome==1.2.0
packaging==21.3
pandas==1.1.5
Pillow==9.3.0
pycparser==2.21
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.6.0
six==1.16.0
smart-open==6.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
wordcloud==1.8.2.2
wsproto==1.2.0
